refactor: modularize server implementation
- Split large index.ts into logical modules
- Created separate handlers for account, Gmail, and calendar operations
- Added type safety with TypeScript interfaces and runtime type guards
- Maintained existing functionality while improving code organization
- All tests passing and Docker build successful
