+++ src/tools/type-guards.ts
@@ -1,0 +1,35 @@
+import { BaseToolArguments, CalendarEventParams } from './types.js';
+
+export function isBaseToolArguments(args: Record<string, unknown>): args is BaseToolArguments {
+  return typeof args.email === 'string';
+}
+
+export function isCalendarEventParams(args: Record<string, unknown>): args is CalendarEventParams {
+  return typeof args.email === 'string' &&
+    (args.query === undefined || typeof args.query === 'string') &&
+    (args.maxResults === undefined || typeof args.maxResults === 'number') &&
+    (args.timeMin === undefined || typeof args.timeMin === 'string') &&
+    (args.timeMax === undefined || typeof args.timeMax === 'string');
+}
+
+export function isEmailEventIdArgs(args: Record<string, unknown>): args is { email: string; eventId: string } {
+  return typeof args.email === 'string' && typeof args.eventId === 'string';
+}
+
+export function assertBaseToolArguments(args: Record<string, unknown>): asserts args is BaseToolArguments {
+  if (!isBaseToolArguments(args)) {
+    throw new Error('Missing required email parameter');
+  }
+}
+
+export function assertCalendarEventParams(args: Record<string, unknown>): asserts args is CalendarEventParams {
+  if (!isCalendarEventParams(args)) {
+    throw new Error('Invalid calendar event parameters');
+  }
+}
+
+export function assertEmailEventIdArgs(args: Record<string, unknown>): asserts args is { email: string; eventId: string } {
+  if (!isEmailEventIdArgs(args)) {
+    throw new Error('Missing required email or eventId parameter');
+  }
+}
